+++ python/cudf/cudf/core/column/timedelta.py
@@ -46,9 +46,6 @@ class TimeDeltaColumn(column.ColumnBase):
         dtype = np.dtype(dtype)
         if data.size % dtype.itemsize:
             raise ValueError("Buffer size must be divisible by element size")
-        if size is None:
-            size = data.size // dtype.itemsize
-            size = size - offset
 
         super().__init__(
             data,